+++ main.js
@@ -12,6 +12,8 @@ import { GameFlowSystem } from "./systems/GameFlowSystem.js";
 import { ResultSystem } from "./systems/ResultSystem.js";
 import { UpgradeSystem } from "./systems/UpgradeSystem.js";
 import { CustomerSystem } from "./systems/CustomerSystem.js";
+import { ExpirationSystem } from "./systems/ExpirationSystem.js";
+import { InventorySystem } from "./systems/InventorySystem.js";
 
 function initGame() {
   UIManager.init();
@@ -19,10 +21,12 @@ function initGame() {
   ResultSystem.init();
   UpgradeSystem.init();
   CustomerSystem.init();
+  ExpirationSystem.init();
+  InventorySystem.init();
 
   EventBus.emit(EVENTS.GAME_INIT);
 
-  console.log("오늘도 정상영업 v1.5 UpgradeSystem 초기화 완료");
+  console.log("오늘도 정상영업 v2.3 상품 재고 및 유통기한 시스템 초기화 완료");
 }
 
 initGame();